+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### documentdb v0.104-0 (Unreleased) ###
+* Add string case support for `$toDate` operator
 
 ### documentdb v0.103-0 (Unreleased) ###
 * Support collation with aggregation and find on sharded collections *[Feature]*